requirements.txt
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
ccxt>=4.0.0  # для биржевых данных
schedule>=1.2.0  # для планировщика
python-telegram-bot>=20.0  # для уведомлений (опционально)
scipy>=1.11.0  # для статистических расчетов